separate AtomicInstant and Instant
